+++ arviz/stats/stats_utils.py
@@ -439,6 +439,11 @@ Pareto k diagnostic values:
    (0.7, 1]   (bad)      {{4:{0}d}} {{8:6.1f}}%
    (1, Inf)   (very bad) {{5:{0}d}} {{9:6.1f}}%
 """
+SCALE_WARNING_FORMAT = """
+The scale is now log by default. Use 'scale' argument or 'stats.ic_scale' rcParam if 
+you rely on a specific value. 
+A higher log-score (or a lower deviance) indicates a model with better predictive 
+accuracy."""
 SCALE_DICT = {"deviance": "deviance", "log": "elpd", "negative_log": "-elpd"}
 
 
@@ -475,6 +480,7 @@ class ELPDData(pd.Series):  # pylint: disable=too-many-ancestors
                 "Count", "Pct.", *[*counts, *(counts / np.sum(counts) * 100)]
             )
             base = "\n".join([base, extended])
+        base = "\n".join([base, SCALE_WARNING_FORMAT])
         return base
 
     def __repr__(self):